Hoist jest mocks above imports (#12)
Attempting to avoid re-transpiling the TS source files when running jest
tests -- that means also skipping the
[`babel-jest`](https://jestjs.io/docs/code-transformation#defaults)-style
transform used by default in jest (and done in ts-jest) to hoist mock
declarations to the top when transforming testing files. Doing the
hoisting by hand instead -- files look weird, but telling myself this is
more more explicit and less magicky anyway. It's not clean - leaving
some hanging comments behind, but that's fine for now.

__Reason no. 431 why I dislike monkey-patch-based mocking.__

Ordinarily, we have 4 instances of the TS files getting
parsed/transpiled in some way:
1. Normal build/type-checking process (can't avoid that)
2. Linting (can't avoid that either)
3. Testing - can potentially avoid this if we run Jest against the
outputs of `#1`
4. Frontend bundling - could also re-use outputs of `#1` here, but with
a caveat (below).

Making Jest use the pre-transpiled outputs does prevent transpiling to
ESM, since Jest's ESM support is unstable. That means we lose
tree-shaking in `#4` when we bundle the CommonJS outputs.
Ideally, would love to move the repo over to `vitest`, since that has
the benefit of supporting ESM and being potentially an order or
magnitude faster in some cases (definitely on cold starts), from
previous experience.