Try limiting fixture loading to only what's needed.
Testing out a hypothesis that imports from `libs/fixtures` have a
non-trivial cost because of the top-level JSON parse and election
hashing that happens on import. That cost is paid more often as packages
get split up (for the build graph caching benefits).

Very rough draft - probably need to add code generation in the other
direction (ts source to json) and store the election hashes so there's a
stronger guarantee of them staying consistent.