[TSBuild] Performance tweaks for the TSBuild worker (#57)
After playing around locally with the `tsc` API options, it seems like
there's a slight boost from holding on to the last `program` created in
each worker and feeding it into the next, to benefit from any reusable
type information from the last run. IIUC, this parallels the behaviour
of the compiler when used with `.tsbuildinfo` files in the `incremental`
build modes.

Also reducing the max worker count to `6` - didn't see benefits locally
past `5` for average changes to the build graph (e.g. re-building all of
`//libs/...`, and nothing past `6` for a full graph re-build. `5` is
probably the better sweet spot for CI , but will see how the next few
runs go.