[Bazel] Upgrade to 7.4.1 and try out cache GC feature (#63)
Bazel's cache garbage collection feature, due for release in v8.0 is
available as an experimental flag in `>=7.4.0` - testing it out to see
if it provides a quicker way to generating a daily clean cache vs the
current approach of running a full build from scratch (which obviously
wouldn't scale well with a growing codebase past a certain point).

GC's set to kick in after `1m` of idle server time - adding a `5m` sleep
after the build to see if the later saved cache size is any different
from the initial restored cache. Should probably add a way to output
related logs after the sleep, if possible.