Update transpilation to enable better sourcemapped test stack traces (#70)
Updating `esbuild` transpilation to write more IDE-friendly filepaths to
the sourcemap.

Was previously following symlinks out of the Bazel output tree,
resulting in sourcemap paths of the form `../../../../path/to/src`. This
enables the `PreserveSymlinks` option and adds an appropriate
`SourceRoot` in order to generate sourcemap paths relative to the repo
root.