[hmpb] Switch QR code from PNG to vector gfx (#86)
Switching to constructing QR codes with vector rectangles to represent
each module, instead of converting to PNG first. Takes up bit more space
in the PDF output, but drops peak memory usage by ~30% and removes the
need to parallelise the QR -> PNG conversion to avoid the bottleneck.
Doesn't seem to be any less performant for readers to render either, at
least for the size of QR generated from the encoded ballots.